Add first working code for sender and receiver
Flush code has been added to sender.c and reload code to receiver.c.
When the current sender is run, it keeps flushing the whole LLC until
stopped. When the current receiver is run, at each user input, it
measures the number of hits and misses for repeated accesses to 8
addresses and prints these numbers.

This whole code is still work in progress and it also contains functions
that have been previously used for experiments, are not used anymore, and
will be removed in the next commit.