Add state and bit detection function to receiver
A state struct with state variables, as well as a function to initialize
it have been added to receiver.c.

The receiver now accepts input flags -w (for the waiting cycles,
defaulting to 40000), -t (for the iterations of the measurements,
defaulting to 100), and -d (for enabling debug prints).

The code to measure the 1 or 0 bit in receiver.c has been moved to
the detect_bit function.

Although the detection function still needs to be tuned up, the
current code is working reliably enough to start working on the
protocol.