Improve code and reliability of communication
The util functions for clflush, string to binary conversion and pow
have been moved to util.c and util.h so that we do not need other
separate files.

The names of static constant variables have been improved.

The code of the sender has been considerably simplified and improved.
The sender now uses a state struct similar to the receiver.

Now the sender sends a sequence "0101010100" before each message
to let the receiver synchronize.

The code of the receiver has been also cleaned up and improved.
The detection function has been changed to be more reliable (all the
parameters are relative to the program arguments, no hardcoded numbers).

The receiver now looks for a "0101" message to synchronize and then
it looks for a "00" message before starting to receive the incoming
message.

TODO:
1) Make the "end string" part be a byte of all 1s, instead of a bit
2) Possibly add error correcting codes.