Optimize the speed and fix a few bugs
The sender now flushes only a few selected addresses that have
cache set index 0.

The receiver probes only a few (> 1) addresses that also have cache
set index 0.

The function to get the cache set now only uses 10 bits, because
the target cache that we are using has 1024 cache sets per slice.

A bug in the ending byte of a received string (receiver side), where
sometimes the zeros were counted starting from the middle of a byte has
been fixed.

The cache_utils.c file has been removed as it is not needed.

A benchmark mode has been added to the sender.c that allows to
measure the sending bitrate.